+++ src/frontend/src/tests/app/AppAuthenticated.test.tsx
@@ -30,7 +30,7 @@ const renderComponent = (path?: string, route?: string) => {
   const RouterWrapper = routerWrapperBuilder({ path, route });
   return render(
     <RouterWrapper>
-      <AppAuthenticated />
+      <AppAuthenticated userId={1} />
     </RouterWrapper>
   );
 };
